chore(server): rename index.ts to server.ts
