+++ apps/server/src/controllers/poiController.ts
@@ -1,5 +1,5 @@
 import type { Request, Response } from "express";
-import { webSocketService } from "../../index";
+import { webSocketService } from "../../server";
 import { handleControllerError } from "../errors/errorHandler";
 import { poiService } from "../services/poiService";
